pyproject.toml
[project]
name = "deep-cam"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "cv2-enumerate-cameras>=1.2.2",
    "gfpgan>=1.3.8",
    "insightface>=0.7.3",
    "numpy>=2.2.6",
    "onnxruntime>=1.22.1",
    "opencv-python>=4.12.0.88",
    "pygrabber>=0.2",
    "scikit-learn>=1.7.1",
    "torch>=2.8.0",
    "torchvision>=0.23.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
